fix: vae -> self.vae

+++ src/diffusers/pipelines/stable_diffusion_xl/pipeline_stable_diffusion_xl.py
@@ -1252,7 +1252,7 @@ class StableDiffusionXLPipeline(
                 latents = latents.to(next(iter(self.vae.post_quant_conv.parameters())).dtype)
 
             if hasattr(self, "is_vae_in_blfoat16") and self.is_vae_in_blfoat16:
-                latents = latents.to(vae.dtype)
+                latents = latents.to(self.vae.dtype)
             print(f"latents dtype: {latents.dtype}")
             image = self.vae.decode(latents / self.vae.config.scaling_factor, return_dict=False)[0]
